fix: body element offset in mobile view
Because the body element has a bult-in offset (through margin) that
shows in almost all modern browsers. Even though we explicitly set
margin: 0 in the app.component.scss, it doesn't render as expected.
This is due to the way Angular encapsulates CSS styles onto a component.
But wait a minute, was this really necessary?
There is a styles.css file in our src folder. This for global styles,
can we just add a class here to style body?
Yes!, but hey at least we learned something here :)