feat: Add the individual Item component
each post (Item) will have its own attributes and so makes sense to
have a component for this. When have real data will need to pass the
item identifier from the story component to the child item component.
for now, will pass the list index as the item ID from story to child.